Extract non-exe logic to commands.
